+++ configs.py
@@ -14,8 +14,6 @@ parser = argparse.ArgumentParser()
 parser.add_argument("--batch", type=int, default=5)
 parser.add_argument("--num_proc", type=int, default=4, help="number of processes")
 parser.add_argument("--n_epochs", type=int, default=10)
-parser.add_argument("--bert_model", type=str, default="bert-base-uncased",
-                    help="default pretrain BERT model")
 parser.add_argument("--n_shards", type=int, help="Number of chunks to split from large dataset",
                     default=8)
 parser.add_argument("--device", type=str, default="default", choices=["default", "cpu", "cuda"],
@@ -40,7 +38,7 @@ else:
 
 ## other args
 args.multi_gpus = torch.cuda.device_count() > 0
-
+args.bert_model         = "bert-base-uncased"
 args.seq_len_ques       = 40
 args.seq_len_para       = 50
 args.seq_len_ans        = 40        # maximum answer length of dataset
@@ -53,15 +51,15 @@ args.d_vocab            = 27154     # Vocab size
 args.dropout            = 0.2
 args.n_layers           = 5
 args.beam_depth         = 5
-args.beam_breadth       = 6
-args.beam_size          = 20
+args.beam_breadth       = 8
+args.beam_size          = 10
 args.beam_ngram_repeat  = 5
 
 
 ###############################
 # Config logging
 ###############################
-logging.basicConfig(filename='app.log', filemode='a+', format='%(asctime)s: %(message)s', datefmt='%b-%d-%Y %H:%M:%S')
+logging.basicConfig(filename='run.log', filemode='a+', format='%(asctime)s: %(message)s', datefmt='%b-%d-%Y %H:%M:%S')
 logging.getLogger().setLevel(logging.INFO)
 
 ###############################
@@ -71,7 +69,8 @@ PATH    = {
     ## Paths associated with Data Reading
     'dataset_para'      : "backup/[SPLIT]/data_[SHARD].csv",
     'processed_contx'   : "backup/proc_contx_[SPLIT].json",
-    'vocab_PGD'         : "backup/vocab_PGD.txt",
+    'vocab'             : "backup/vocab.txt",
     'saved_model'       : "backup/model.pt",
-    'saved_chkpoint'    : "backup/chkpoint.pth.tar"
+    'saved_chkpoint'    : "backup/chkpoint.pth.tar",
+    'eval_result'       : "backup/metrics.json"
 }